July 2020

+++ scripts/copy-pstweetchat.ps1
@@ -2,6 +2,12 @@
 [cmdletbinding(SupportsShouldProcess)]
 Param()
 
-Get-ChildItem C:\Users\Jeff\downloads\PSTweetchat* -ov csv | Sort-Object lastwritetime | Select-Object -last 1 | Copy-Item -Destination $PSScriptRoot\pstweetchat.csv -PassThru
+$dl = Get-ChildItem C:\Users\Jeff\downloads\PSTweetchat* -ov csv | Sort-Object lastwritetime | Select-Object -last 1 | Copy-Item -Destination $PSScriptRoot\pstweetchat.csv -PassThru
 
-$csv | Remove-Item
+if ($dl) {
+    $dl
+    $csv | Remove-Item
+}
+else {
+    Write-Warning "No CSV files found"
+}
